fix get all product

+++ lib/features/admin/add_products/data/models/get_all_product_response.dart
@@ -45,7 +45,7 @@ class ProductGetAllModel {
 
   final String? id;
   final String? title;
-  final int? price;
+  final double? price;
   final List<String>? images;
   final String? description;
   final CategoryProductModel? category;
